Simplify AWS environment by removing unused availablity zones
When trying to deploy this infrastructure to ap-south-1, one of its
three availability zones did not have client VPN capabilities.
Therefore we are simplifying the code to only use one AZ.  Though it is
a good practice to use multiple AZs to increase fault tolerance, this
training code is not fully utilizing that design.